Reconciled student views, hid capital gains tax
I commented it out for now since it doesn't exist in the spreadsheet yet, but we can put it in later once it does. Technically the same from the student side since it only shows up on admin side, but hidden for now since it's just wrong (it showed Transfers Out instead).